+++ link_test.go
@@ -17,6 +17,7 @@ import (
 func TestLinkFlowThatNeedsToReplenishCredits(t *testing.T) {
 	for times := 0; times < 100; times++ {
 		l := newTestLink(t)
+		l.l.linkCredit = 2
 		go l.mux()
 
 		err := l.DrainCredit(context.Background())
@@ -25,20 +26,11 @@ func TestLinkFlowThatNeedsToReplenishCredits(t *testing.T) {
 		err = l.IssueCredit(1)
 		require.Error(t, err, "issueCredit can only be used with receiver links using manual credit management")
 
-		// and flow goes through the non-manual credit path
-		require.EqualValues(t, 0, l.l.linkCredit, "No link credits have been added")
-
 		// we've consumed half of the maximum credit we're allowed to have - reflow!
-		l.maxCredit = 2
 		l.l.linkCredit = 1
 		l.unsettledMessages = map[string]struct{}{}
 
-		select {
-		case l.receiverReady <- struct{}{}:
-			// woke up mux
-		default:
-			t.Fatal("failed to wake up mux")
-		}
+		require.NoError(t, l.onSettlement(1))
 
 		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 	Loop:
@@ -82,7 +74,6 @@ func TestLinkFlowWithZeroCredits(t *testing.T) {
 	// and flow goes through the non-manual credit path
 	require.EqualValues(t, 0, l.l.linkCredit, "No link credits have been added")
 
-	l.maxCredit = 2
 	l.l.linkCredit = 0
 	l.unsettledMessages = map[string]struct{}{
 		"hello":  {},
@@ -116,7 +107,6 @@ func TestLinkFlowWithManualCreditor(t *testing.T) {
 	l := newTestLink(t)
 	l.autoSendFlow = false
 	l.l.linkCredit = 1
-	l.maxCredit = 1000
 	go l.mux()
 	defer close(l.l.close)
 
@@ -132,17 +122,6 @@ func TestLinkFlowWithManualCreditor(t *testing.T) {
 	default:
 		require.Fail(t, fmt.Sprintf("Unexpected frame was transferred: %+v", txFrame))
 	}
-}
-
-func TestLinkFlowWithManualCreditorTooManyCredits(t *testing.T) {
-	l := newTestLink(t)
-	l.autoSendFlow = false
-	l.l.linkCredit = 1
-	l.maxCredit = 100
-	go l.mux()
-	defer close(l.l.close)
-
-	require.Error(t, l.IssueCredit(100))
 }
 
 func TestLinkFlowWithDrain(t *testing.T) {
@@ -244,9 +223,10 @@ func newTestLink(t *testing.T) *Receiver {
 			rx:    make(chan frames.FrameBody, 100),
 			close: make(chan struct{}),
 		},
-		autoSendFlow:  true,
-		inFlight:      inFlight{},
-		receiverReady: make(chan struct{}, 1),
+		autoSendFlow:    true,
+		inFlight:        inFlight{},
+		settlementCount: make(chan uint32),
+		receiverReady:   make(chan struct{}, 1),
 	}
 
 	l.messagesQ = queue.NewHolder(queue.New[Message](100))